Move HotRestartCacheHandler to its own file.
